pyproject.toml
[project]
name = "cash-flow-planner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jinja2>=3.1.5",
    "py-moneyed>=3.0",
    "streamlit>=1.42.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "pre-commit>=4.1.0",
    "pytest>=8.3.4",
]


[tool.pytest.ini_options]
pythonpath = [
    "test"
]
